Rename Inpage component filename.
