+++ app/assets/scripts/components/auth/index.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import { environment, apiUrl, baseUrl } from '../../config';
-import { Inpage, InpageBody } from '../common/Inpage';
+import { Inpage, InpageBody } from '../common/inpage';
 import * as authActions from '../../redux/actions/auth';
 import { PropTypes as T } from 'prop-types';
 import { showGlobalLoading, hideGlobalLoading } from '../common/global-loading';
