+++ app/assets/scripts/components/common/inpage.js
